Fix bug in AdvInd propagation
When handling main array updates, if there was an update to an index
that was not covered by the just-parsed updates to indexing arrays, then
a function to break the index into its main offset and "sub array index"
(called `get_mapped_index`) would segfault if the indexing arrays were
not 1d. This went unnoticed by the fixes in #112 because we forgot to
test this special kind of update on the main array.