Fix two bugs in advanced indexing propagation
First one was just a simple logic error in which method should be called
when reverting the reverse offset map. Second one was basically covered
by an old TODO: to ensure all updates to the parent array are reflected
in the final advanced indexing output, we need to track the offset
indices that were modified (not just the offsets themselves) as new
offsets may be added that duplicate previous unchanged offsets.