Fix bug in basic indexing propagation
During basic indexing propagation when the predecessor array is
statically sized or both the start and stop are positive, we do a basic
translation of indices of the updates and otherwise pass them through.
We also calculate the start and stop of the "domain" of the basic
indexing operation in terms of indices on the predecessor array, and
then exclude updates that are outside of this domain. The stop of this
domain was calculated by subtracting the pointers returned by `end()`
and `begin()`. However, if the start of the domain is not zero, this
stop index will not be correct, and thus some updates may get excluded
that should not be. Fixed by subtracting the predecessor array's
`begin()` from the basic indexing node's `end()` to get the true stop
index.