Add IsInNode/Symbol
Initial minimal implementation of numpy.isin() using lookup tables.
Fixes #76. Given that this fixes an old issue, I wanted to check which
functionality from numpy.isin() we should port over for this node.
Should this node include any of the following?

1. bool assume_unique (kwarg)
2. bool invert (kwarg)
3. kind {None, `sort`, `table`} (kwarg)

Regarding 2, I believe this could just be done with the EqualNode.
For example:

>>> from dwave.optimization.model import Model
>>> from dwave.optimization.mathematical import isin
>>> model = Model()
>>> model.states.resize(1)
>>> element = model.constant([1.3, -1.0, 3.0])
>>> test_elements = model.constant([3.0, -1.1, -2.0, 4.1, -1.0])
>>> contains = isin(element, test_elements)
>>> inverted_contains = contains == [0.0, 0.0, 0.0]
>>> with model.lock():
...     print(contains.state(0))
...     print(inverted_contains.state(0))
[0.0, 1.0, 1.0]
[1.0, 0.0, 0.0]

Regarding 3, given we already have an ArgSortNode, it seems reasonable
that we could pass this node a bool like `sorted` which could indicate
whether its predecessors are ArgSortNodes. We could implement methods
to deal with these cases and call them or the current lookup table
methods from `IsInNode::initialize_state`` and `IsInNode::propagate``.