Fix potential issue with expression input init in AccumulateZipNode
Previously AccumulateZip would initialize the inputs in an expression
their `min()` value, but since it is common that we have unbounded
inputs (i.e. `min()` returns `-inf`) this could result in NaNs being
created/propagated in the expression during evaluation, causing other
issues. This is fixed by just initializing the inputs close to 0.